warn superuser deletion
git-svn-id: svn://svn.pgadmin.org/trunk@3452 a7884b65-44f6-0310-8a51-81a127f17b15

+++ src/schema/pgUser.cpp
@@ -35,6 +35,15 @@ pgUser::~pgUser()
 
 bool pgUser::DropObject(wxFrame *frame, wxTreeCtrl *browser)
 {
+    if (GetUpdateCatalog())
+    {
+        wxMessageDialog dlg(frame, 
+            _("Deleting a superuser might result in unwanted behaviour (e.g. when restoring the database).\nAre you sure?"),
+            _("Confirm superuser deletion"),
+                     wxICON_EXCLAMATION | wxYES_NO |wxNO_DEFAULT);
+        if (dlg.ShowModal() != wxID_YES)
+            return false;
+    }
     return server->ExecuteVoid(wxT("DROP USER ") + GetQuotedFullIdentifier() + wxT(";"));
 }
